+++ actions/authActions.js
@@ -164,20 +164,15 @@ export const AUTH_REQUEST = 'AUTH_REQUEST';
 export const AUTH_SUCCESS = 'AUTH_SUCCESS';
 // export const AUTH_FAIL = 'AUTH_FAIL';
 
-
-
 const authRequest = () => {
   return {
     type: AUTH_REQUEST
   }
 }
 
-const authSuccess = (username) => {
+export const authSuccess = (username) => {
   return {
     type: AUTH_SUCCESS,
     username: username
   }
 }
-
-
-
