--- reducers/index.js
@@ -1,2 +1,0 @@
-// export beerReducer from './beerReducers'
-// export authReducer from './authReducers'
